app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.4.2
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.0
tiktoken==0.5.1
beautifulsoup4>=4.12.0
opencv-python>=4.8.0
tensorflow>=2.13.0
imagehash>=4.3.1
python-whois>=0.8.0
tldextract>=3.4.0
cssutils>=2.6.0
selenium>=4.10.0
webdriver-manager>=3.8.0
Pillow>=10.0.0
numpy>=1.24.0
difflib>=3.5.0

# Optional: For screenshot and web interaction
chromedriver-binary>=119.0.0